+++ src/App.vue
@@ -1,9 +1,15 @@
 <script setup>
 import { ref } from 'vue';
 import { storyData } from '@/stores/story'
+import { createToaster } from "@meforma/vue-toaster";
 
 const story = storyData()
 const installPrompt = ref(null)
+const toaster = createToaster({
+  position : 'top',
+  duration : 2000,
+  dismissible : true
+});
 
 window.addEventListener("beforeinstallprompt", e => {
   if (story.askInstall) {
@@ -24,6 +30,128 @@ const neverIntall = () => {
   installPrompt.value = null
   story.askInstall = false
 }
+
+function cleaningStoryFormats(data){
+  let result = {}
+  if (data.split(`"setup": function(){`).length > 1) {
+    result = data.split(`"setup": function(){`)[0].slice(0, -1) + '}'
+    result = result.slice(19); 
+    result = JSON.parse(result)
+  } else {
+    result = data.substring(data.length - 1) == ";" ? data.slice(0, -2) : data.slice(0, -1)
+    result = result.slice(19); 
+    result = JSON.parse(result)
+  }
+  return result
+}
+
+const chapbook_1 = "https://raw.githubusercontent.com/klembot/twinejs/dc8aaeae7c3c279e12f04f74093e315027b8e15d/public/story-formats/chapbook-1.2.3/format.js"
+const chapbook_2 = "https://raw.githubusercontent.com/klembot/twinejs/911d18bc4994f1648137da4cacb66fced9086f6f/public/story-formats/chapbook-2.2.0/format.js"
+const harlowe_1 = "https://raw.githubusercontent.com/klembot/twinejs/28074261074cf22e46779c0efc2442afe015b74c/public/story-formats/harlowe-2.1.0/format.js"
+const harlowe_2 = "https://raw.githubusercontent.com/klembot/twinejs/28074261074cf22e46779c0efc2442afe015b74c/public/story-formats/harlowe-2.1.0/format.js"
+const harlowe_3 = "https://raw.githubusercontent.com/klembot/twinejs/60c3cadc308ec9ceafb0406c5c1ef58eaa828d2e/public/story-formats/harlowe-3.3.9/format.js"
+const snowman_1 = "https://raw.githubusercontent.com/klembot/twinejs/dd79dc7073115b1acfc82f25dace9838c241e056/public/story-formats/snowman-1.4.0/format.js"
+const snowman_2 = "https://raw.githubusercontent.com/klembot/twinejs/dd79dc7073115b1acfc82f25dace9838c241e056/public/story-formats/snowman-2.0.2/format.js"
+// const sugarcube_1 = "https://raw.githubusercontent.com/klembot/twinejs/c5a2c90b86bb1d3d50169b0fe3988d3434c23127/public/story-formats/sugarcube-1.0.35/format.js"
+const sugarcube_2 = "https://raw.githubusercontent.com/klembot/twinejs/b2f8717bde8f0ef16ca78d9b99b6139c3bd05321/public/story-formats/sugarcube-2.37.3/format.js"
+story.storyFormatList.forEach(storyList => {
+  if (storyList.val == 'chapbook-1' && !storyList.finishGet) {
+    fetch(chapbook_1)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'chapbook-2' && !storyList.finishGet) {
+    fetch(chapbook_2)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'harlowe-1' && !storyList.finishGet) {
+    fetch(harlowe_1)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'harlowe-2' && !storyList.finishGet) {
+    fetch(harlowe_2)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'harlowe-3' && !storyList.finishGet) {
+    fetch(harlowe_3)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'snowman-1' && !storyList.finishGet) {
+    fetch(snowman_1)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'snowman-2' && !storyList.finishGet) {
+    fetch(snowman_2)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+  if (storyList.val == 'sugarcube-2' && !storyList.finishGet) {
+    fetch(sugarcube_2)
+      .then(response => response.text())
+      .then(data => {
+        let theResult = cleaningStoryFormats(data)
+        storyList.finishGet = true
+        storyList.data = theResult
+      }).catch((error) => {
+        toaster.error(`Something wrong... Check your internet connection...`)
+        console.log('error', error)
+      });
+  }
+})
 
 story.innerHeight = window.innerHeight
 </script>